Merge design-doc into stage (#6)
* Create design doc

* Add initial details to design doc

* Insert wireframe section and link into design doc

I finished the wireframe on figma and put the link into the design doc. Everyone
will need to make a figma account to access the wireframe. I left comments throughout
the wireframe for suggested flow of the feature. These comments are obviously open for
debate and I suspect we will need a few more features not present in the wireframe. This
serves as a baseline to get us started on the project and headed in the right direction.

* Write user stories

Write stories for Sally Student and Amy Ambassador. Should Sally be
able to reserve for the future or just checkout based on what is
currently available?

* Add data security and privacy concerns to design-doc

* Add another capabaility for Amy Ambassador in design-doc

* Add tech implementation #1-#3

* Update Amy Ambassador Story

Add to the story that Amy should be able to view the checkout history of each
story.

* Add possible API routes needed

* Add change proposed in PR review

---------

Co-authored-by: David Sprague <dsprague@unc.edu>
Co-authored-by: Jacob Brown <jsbrown@unc.edu>
Co-authored-by: nicholasmountain11 <nmountain@unc.edu>
Co-authored-by: Jacob Brown <97564414+jacbro2021@users.noreply.github.com>
Co-authored-by: David Sprague <97635001+davids122@users.noreply.github.com>