9 backend models and entities (#21)
* Create model and entity for equipment

Made Pydantic model for equipment and corresponding EquipmentEntity.
Added field in the user model and entity for current checkout

* Added field to model and entity for image

Added a field to model and entity for image and updated the toModel, fromModel, and
update methods on the entity. Also added the condition field to the entity and made
the necessary conditions to the above methods for this change as well.

Changed the update method on Equipment Entity to throw an error if the equipment_id
of the passed in model does not match that of the entity. This should save us from
potential headaches in the future with mismatched updates of entities from models.

* Create equipment service and 'EquipmentNotFoundException' (#17)

* Create equipment service and EquipmentNotFound Exception

Implemented the equipment service with basic
methods for checking equipment in and out. The
service will need to be altered in future sprints
as we create tables to store notes about
individual items and as we implement permissions.

The EquipmentNotFound Exception is needed to raise
an error when equipment is checked in/out that
does not exist in the db.

* Add 'EquipmentService' method for updating condition

* Edit equipment service to combine existing methods into update method

Methods for checking equipment out/in and updating equipment condition
combined into one update method for ease of use. It will be the
responsibility of the API or frontend to edit a model with needed
changes before sending to the service and ultimately updating the db.

methods were also added for inserting equipment into the db and deleting
equipment from the db as deletion will be necessary when equipment is no
longer usable and new equipment will periodically need to be added.

* Change entities so that tests can run

Give EquipmentEntity a tablename
Remove current_equipment field from userEntity so that tests would run.
Will need to figure out why that was not working.

* Write first test for equipment service

Getting the tests to actually run was much more difficult than I hoped,
but I think I gained some important knowledge along the way.
We can retrieve a list of equipment items from the database, more
rigorous tests need to be written, but it is a good start.

* Add Comments

* Write test for equipment service update method

The update test does not pass yet, to be figured out tomorrow.

* Write tests for equipment service (#20)

* Change entities so that tests can run

Give EquipmentEntity a tablename
Remove current_equipment field from userEntity so that tests would run.
Will need to figure out why that was not working.

* Write first test for equipment service

Getting the tests to actually run was much more difficult than I hoped,
but I think I gained some important knowledge along the way.
We can retrieve a list of equipment items from the database, more
rigorous tests need to be written, but it is a good start.

* Add Comments

* Write test for equipment service update method

The update test does not pass yet, to be figured out tomorrow.

* Add additional test for getAll() method

---------

Co-authored-by: Jacob Brown <jsbrown@unc.edu>
Co-authored-by: Jacob Brown <97564414+jacbro2021@users.noreply.github.com>